Deprecate old schema and refactor with new schema
