+++ .vscode/settings.json
@@ -1,0 +1,3 @@
+{
+    "java.jdt.ls.java.home": ""
+}
