+++ src/main/java/com/crowdin/client/reports/model/PreTranslateEfficiencyGenerateReportRequest.java
@@ -1,36 +1,12 @@
 package com.crowdin.client.reports.model;
 
-import lombok.Data;
-import lombok.EqualsAndHashCode;
-
-import java.util.Date;
-import java.util.List;
-
-@Data
-@EqualsAndHashCode(callSuper = true)
-public class PreTranslateEfficiencyGenerateReportRequest extends GenerateReportRequest {
-    private String name = "pre-translate-efficiency";
-    private Schema schema;
-
-    @Data
-    public static class Schema {
-        private Unit unit = Unit.WORDS;
-        private ReportsFormat format = ReportsFormat.XLSX;
-        private List<String> postEditingCategories;
-    }
-
-    @Data
-    @EqualsAndHashCode(callSuper = true)
-    public static class GeneralSchema extends Schema {
-        private String languageId;
-        private Date dateFrom;
-        private Date dateTo;
-    }
-
-    @Data
-    @EqualsAndHashCode(callSuper = true)
-    public static class ByTaskSchema extends Schema {
-        private Long taskId;
-    }
-
+/**
+ * @deprecated Use {@link PreTranslateAccuracyGenerateReportRequest} instead
+ */
+@Deprecated
+public class PreTranslateEfficiencyGenerateReportRequest extends PreTranslateAccuracyGenerateReportRequest {
+	public PreTranslateEfficiencyGenerateReportRequest() {
+		super();
+		this.setName("pre-translate-efficiency");
+	}
 }
